split up homepage into mulitple compoments to make things easier, added search bar to page and project cards too
